Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
youtube-transcript-api==0.6.2
google-generativeai==0.8.3
faiss-cpu==1.7.4
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
rpunct==1.0.2
